controller: move custom logic to another file
This makes it easier to upgrade kubebuilder, as kubebuilder upgrade
requires to manually update the file or regenerate them with the CLI and
integrating our own change in them.
